% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_dt.R
\name{count_dt}
\alias{count_dt}
\alias{add_count_dt}
\title{Count observations by group}
\usage{
count_dt(data, ..., weight = NULL, sort = FALSE, name = "n")

add_count_dt(data, ..., weight = NULL, sort = FALSE, name = "n")
}
\arguments{
\item{data}{data.table/data.frame data.frame will be automatically converted
to data.table.}

\item{...}{variables to group by.}

\item{weight}{optional. Unquoted variable name. If provided result will be the sum of this variable by groups.}

\item{sort}{logical. If TRUE result will be sorted in desending order by resulting variable.}

\item{name}{character. Name of resulting variable.}
}
\value{
data.table
}
\description{
Analogous function for \code{count} and \code{add_count} in \pkg{dplyr},
wrapped directly from \pkg{maditr}.
}
\examples{
iris \%>\% count_dt(Species)
iris \%>\% add_count_dt(Species)
}
\seealso{
\code{\link[dplyr]{count}},\code{\link[maditr]{dt_count}}
}
