% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dt.R
\name{filter_dt}
\alias{filter_dt}
\title{Filter entries in data.frame}
\usage{
filter_dt(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}
}
\value{
data.table
}
\description{
Analogous function for \code{filter} in \pkg{dplyr}.
}
\examples{
 # filter by condition
 iris \%>\% filter_dt(Sepal.Length == max(Sepal.Length))

 # filter by compound condition
 mtcars \%>\% filter_dt(am==0,cyl == 4)
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[maditr]{dt_filter}}
}
