% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide.R
\name{wider_dt}
\alias{wider_dt}
\alias{longer_dt}
\title{Convert data between wide and long forms}
\usage{
wider_dt(data, group, class_to_spread, value_to_spread = NULL, fill = NA)

longer_dt(
  data,
  group,
  gather_class = "class",
  gather_value = "value",
  na.rm = FALSE
)
}
\arguments{
\item{data}{data.table}

\item{group}{The unchanged group in the transformation}

\item{class_to_spread}{The class of variable to spread}

\item{value_to_spread}{The value of variable to spread. If missing, will guess automatically.}

\item{fill}{Value with which to fill missing cells. Default uses \code{NA}.}

\item{gather_class}{The column name of gathered class.}

\item{gather_value}{The column name of gathered value.}

\item{na.rm}{If TRUE, entries with NA values will be removed from the data.table.}
}
\value{
data.table
}
\description{
Analogous function for \code{dcast} and \code{melt} in \pkg{data.table}
}
\details{
In the \code{nest_dt}, the data would be nested to a column named `ndt`,
 which is short for nested data.table.
}
\examples{
stocks = data.frame(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4)
)

stocks \%>\%
  longer_dt(time) -> longer_stocks

longer_stocks

longer_stocks \%>\%
  wider_dt(time,class)

longer_stocks \%>\%
  mutate_dt(one = 1) \%>\%
  wider_dt(time,class,one)

}
\seealso{
\code{\link[data.table]{dcast}},\code{\link[data.table]{melt}},\code{\link[maditr]{dcast}}

\code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{pivot_wider}}
}
