% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_dt.R
\name{mutate_dt}
\alias{mutate_dt}
\title{Mutate columns in data.frame}
\usage{
mutate_dt(data, ..., by)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}

\item{by}{unquoted name of grouping variable of list of unquoted names of
grouping variables. For details see \link[data.table]{data.table}}
}
\value{
data.table
}
\description{
Analogous function for \code{mutate} in \pkg{dplyr}.This is a little different from
\code{\link[maditr]{dt_mutate}},because it prohibit the in place replacement.
}
\examples{

iris \%>\% mutate_dt(one = 1)
mtcars  \%>\%
  head(3) \%>\%
  mutate_dt(
    cyl2 = cyl * 2,
    cyl4 = cyl2 * 2
  )

# window functions are useful for grouped mutates
iris \%>\%
  mutate_dt(
    rank = rank(-Sepal.Length, ties.method = "min"),
    keyby = Species)
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[maditr]{dt_mutate}}
}
