% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_dt.R
\name{nest_dt}
\alias{nest_dt}
\alias{unnest_dt}
\title{Nest and unnest}
\usage{
nest_dt(data, ...)

unnest_dt(data, col)
}
\arguments{
\item{data}{data.table, nested or unnested}

\item{...}{The variable for nest group}

\item{col}{The variable of nested data to be unnest}
}
\value{
data.table, nested or unnested
}
\description{
Analogous function for \code{nest} and \code{unnest} in \pkg{tidyr}
}
\details{
In the \code{nest_dt}, the data would be nested to a column named `ndt`,
 which is short for nested data.table.
}
\examples{
iris \%>\% nest_dt(Species) -> nested_iris
nested_iris

nested_iris \%>\% unnest_dt(ndt) -> unnested_iris
unnested_iris
}
\references{
https://www.r-bloggers.com/much-faster-unnesting-with-data-table/

https://stackoverflow.com/questions/25430986/create-nested-data-tables-by-collapsing-rows-into-new-data-tables
}
\seealso{
\code{\link[tidyr]{nest}}
}
