% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dt.R
\name{select_dt}
\alias{select_dt}
\title{Select column from data.frame}
\usage{
select_dt(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}
}
\value{
data.table
}
\description{
Analogous function for \code{select} in \pkg{dplyr}.
}
\examples{
  iris \%>\% select_dt(1,3,4)
  iris \%>\% select_dt(2:3)
  iris \%>\% select_dt(Sepal.Length)
  iris \%>\% select_dt(Sepal.Length,Species)
  iris \%>\% select_dt(-Sepal.Length,-Species)
  iris \%>\% select_dt(startsWith(colnames(.),"Petal"))
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[maditr]{dt_mutate}}
}
