% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dt.R
\name{summarise_dt}
\alias{summarise_dt}
\alias{summarize_dt}
\alias{summarise_all_dt}
\alias{summarize_all_dt}
\title{Summarise columns to single values}
\usage{
summarise_dt(data, ..., by, keyby, fun = NULL)

summarize_dt(data, ..., by, keyby, fun = NULL)

summarise_all_dt(data, fun, by, keyby)

summarize_all_dt(data, fun, by, keyby)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}

\item{by}{unquoted name of grouping variable of list of unquoted names of
grouping variables. For details see \link[data.table]{data.table}}

\item{keyby}{Same as \code{by}, but with an additional \code{setkey()} run on the by
columns of the result, for convenience. It is common practice to use
'keyby=' routinely when you wish the result to be sorted. For details see
\link[data.table]{data.table}.}

\item{fun}{function which will be applied to all variables in
\code{dt_summarize} and \code{dt_summarize_all}.}
}
\value{
data.table
}
\description{
Analogous function for \code{summarise} and \code{summarise_all} in \pkg{dplyr}.
}
\examples{
iris \%>\% summarise_dt(avg = mean(Sepal.Length))
iris \%>\% summarise_dt(avg = mean(Sepal.Length),by = Species)
mtcars \%>\% summarise_dt(avg = mean(hp),by = .(cyl,vs))

# the data.table way
mtcars \%>\% summarise_dt(cyl_n = .N, by = .(cyl, vs)) # `.`` is short for list

iris \%>\%summarise_all_dt(fun = mean,by = Species)


}
\seealso{
\code{\link[dplyr]{summarise}},\code{\link[maditr]{dt_summarise}},
  \code{\link[dplyr]{summarise_all}},\code{\link[maditr]{dt_summarise_all}}
}
