% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topn_dt.R
\name{top_n_dt}
\alias{top_n_dt}
\alias{top_frac_dt}
\title{Select top (or bottom) n rows (by value)}
\usage{
top_n_dt(data, n, wt)

top_frac_dt(data, n, wt)
}
\arguments{
\item{data}{data.frame}

\item{n}{If \code{n} is positive, selects the top rows. If negative, selects the bottom rows.}

\item{wt}{The numeric variable to use for ordering.}
}
\value{
data.table
}
\description{
Analogous function for \code{top_n} and \code{top_frac} in \pkg{dplyr}, but with a different API.
}
\examples{
iris \%>\% topn_dt(Sepal.Length,10)
iris \%>\% topn_dt(Sepal.Length,-10)
iris \%>\% topfrac_dt(Sepal.Length,.1)
iris \%>\% topfrac_dt(Sepal.Length,-.1)

}
\seealso{
\code{\link[dplyr]{top_n}},\code{\link[maditr]{dt_top_n}}
}
